tests_install/requirements.txt
distro
allure-pytest
beautifulsoup4
kiwisolver
matplotlib
netaddr
numpy
psutil
pytest
pytest-html
pytest-json
enum34
